fix: Default Menu Tree in role. (#421)
* fix: Default Menu Tree in role.

When a role does not have a menu tree established, the company's menu tree is used. If the company does not have one, it sends null and the menu will never load.

This is corrected by searching for the default menu or, if there is no default menu, the one with the lowest ID associated with the role's company or the system company.

* change tree type to `MM`.